client/requirements.txt
click>=8.1
requests>=2.31
tabulate>=0.9
pyyaml>=6.0
